- Adding tracker for changes to project - Fixing bugs on tab switching - Fixing canvas is undefined bug on shader editor
