+++ frontend/components/view/Views.svelte
@@ -6,6 +6,8 @@
     import removeTab from "./utils/remove-tab";
     import addTab from "./utils/add-tab";
     import ResizableBar from "../resizable/ResizableBar.svelte";
+    import SettingsStore from "../../editor/stores/SettingsStore";
+    import {onDestroy} from "svelte";
 
     export let resizePosition
     export let orientation
@@ -42,6 +44,7 @@
             Object.assign(ref.style, obj)
         }
     }
+
 
 </script>
