+++ frontend/components/window-frame/GlobalOptions.svelte
