+++ frontend/components/window-frame/components/ActionHistory.svelte
