+++ frontend/components/window-frame/components/Console.svelte
