+++ frontend/components/window-frame/components/LevelSelector.svelte
